[add]外部APIを利用する為にServiceクラスを作成
NEWSAPIからデータを取得する
